[Xposed] Support gzipped odex files
There are a couple of unmerged AOSP contributions by Intel:
https://android-review.googlesource.com/#/c/145091/
https://android-review.googlesource.com/#/c/149116/
https://android-review.googlesource.com/#/c/149243/

These changes would add support for gzipped odex files, which would be inflated
by patchoat into the Dalvik cache. CM and possibly other ROMs have includeded
these changes.

In Xposed, patchoat isn't used, so dex2oat has to be able to inflate these files
temporarily before they can be recompiled. This commit is based on the patchoat
changes above. It uses the swap file (which seems to be always available when
dex2oat is called from installd) for inflating.

There's one catch: The second change modifies installd in a way that it always
calls patchoat for .gz files - even if the initial request was meant to call
dex2oat. In order to avoid this, the flash script should rename those files
to .gz.xposed, which will then be handled by this commit.